main.py
import asyncio, os

from dotenv import load_dotenv
from livekit.agents import AutoSubscribe, JobContext, WorkerOptions, cli, llm
from livekit.agents.llm import ChatContext, ChatMessage
from livekit.agents.voice_assistant import VoiceAssistant
from livekit.plugins import openai, silero, deepgram, cartesia

load_dotenv()

example_flow = """
Hi, I'm here to help estimate your monthly net income. First, are you employed by a company or self-employed?

User: I'm self-employed.

Assistant: Great, thanks for letting me know. Do you have a rough idea of how much you earn on average each month?

User: Not really, it changes a lot.

Assistant: That's okay. Let's break it down. How much did you earn last month?

User: Last month, I made about $6,500.

Assistant: Got it. And what about the month before that?

User: That was around $4,000.

Assistant: Thanks. And the month before that?

User: That one was about $5,500.

Assistant: Great, based on the last three months, your average monthly income is around $5,333. Now, let’s figure out your expenses. Do you know how much you typically spend each month?

User: Not really.

Assistant: Alright, let’s break it down. How much do you spend on housing, like rent or mortgage?

User: $1,500.

Assistant: And for essentials like food, utilities, and transportation?

User: About $1,000.

Assistant: Got it. Do you have any regular business expenses, like supplies or services?

User: Yes, around $800 a month.

Assistant: Thanks for the details. Your total monthly expenses are about $3,300. Based on your income and expenses, your estimated monthly net income is about $2,033.

User: That sounds right.

Assistant: Glad I could help. Let me know if you need anything else.
"""


async def entrypoint(ctx: JobContext):
    chat_context = ChatContext(
        messages=[
            ChatMessage(
                role="system",
                content=(
                    # "You are an employee of a lending company. Your interface with the users is voice."
                    # "Your job is to ask questions to a client to estimate their monthly net income. Net income is total income - expenses."
                    # "Wait for their answer then ask the next question."
                    # "Keep your responses short and concise and avoid usage of complicated and unpronoucable words."
                    # "When doing math, never say the operations and only give the final answer."
                    "You are a loan agent for a lending company and you talk to borrowers via voice. Your task is to ask the borrower a series of questions to estimate their net monthly income. Net income is calculated as total income minus expenses."
                    "Ask clear, simple questions one at a time."
                    "Wait for the user’s response before moving to the next question."
                    "Speak in short, concise sentences. Avoid complex words or jargon."
                    "When providing any calculation, only state the result—no need to explain the math steps."
                    "Keep the conversation polite and professional."
                    "Keep the conversation smooth, quick, and to the point, aiming for a friendly and professional tone throughout."
                    f"Here is an example flow: {example_flow}"
                ),
            )
        ]
    )
    await ctx.connect(auto_subscribe=AutoSubscribe.AUDIO_ONLY)

    assistant = VoiceAssistant(
        vad=silero.VAD.load(),
        stt=deepgram.STT(api_key=os.getenv("DEEPGRAM_API_KEY")),
        llm=openai.LLM.with_groq(
            api_key=os.getenv("GROQ_API_KEY"),
            model="llama-3.1-70b-versatile",
        ),
        tts=cartesia.TTS(api_key=os.getenv("CARTESIA_API_KEY")),
        chat_ctx=chat_context,
    )
    assistant.start(ctx.room)

    await asyncio.sleep(1)
    await assistant.say(
        "Hey, I am an assistant from airblock.", allow_interruptions=True
    )


if __name__ == "__main__":
    cli.run_app(WorkerOptions(entrypoint_fnc=entrypoint))
