requirements.txt
aiodns==3.2.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2
async-timeout==4.0.3
attrs==24.2.0
av==13.1.0
certifi==2024.8.30
cffi==1.17.1
click==8.1.7
coloredlogs==15.0.1
distro==1.9.0
exceptiongroup==1.2.2
flatbuffers==24.3.25
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
humanfriendly==10.0
idna==3.10
jiter==0.6.1
livekit==0.17.5
livekit-agents==0.10.1
livekit-api==0.7.1
livekit-plugins-cartesia==0.4.2
livekit-plugins-deepgram==0.6.8
livekit-plugins-elevenlabs==0.7.5
livekit-plugins-openai==0.10.3
livekit-plugins-silero==0.7.2
livekit-protocol==0.6.0
mpmath==1.3.0
multidict==6.1.0
numpy==1.26.4
onnxruntime==1.19.2
openai==1.51.2
packaging==24.1
pillow==10.3.0
propcache==0.2.0
protobuf==5.28.2
psutil==5.9.8
pycares==4.4.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
python-dotenv==1.0.1
sniffio==1.3.1
sympy==1.13.3
tqdm==4.66.5
types-protobuf==4.25.0.20240417
typing_extensions==4.12.2
watchfiles==0.24.0
yarl==1.15.2
